ServerAPIService: Split off APIv2-specific stuff separately
